feat(categories): redesign categories page with accordion layout and search
Replace the card grid layout with an accordion-style list for better tag
discoverability and overview. Key UX improvements:

- Add global search bar that filters both categories and tags with highlighting
- Add stats badges showing total categories and tags count
- Switch from card grid to expandable accordion rows with tag count badges
- Add "Toggle All" expand/collapse button for full taxonomy overview
- Categories auto-expand when search matches a tag inside them
- Extract emoji data and icon picker into dedicated component files
- Use shared Modal component for all modals (icon picker, create, relocate)
- Rename category via dedicated pencil icon button instead of click-on-name
