+++ frontend/src/components/categories/emojiData.ts
@@ -1,0 +1,576 @@
+// [emoji, searchKeywords] — keywords are space-separated for fast substring matching
+export const EMOJI_DATA: [string, string][] = [
+  // 💵 Money & Finance
+  ["💰", "money bag gold finance rich wealth"],
+  ["💵", "dollar money cash bill currency"],
+  ["💴", "yen money japan currency"],
+  ["💶", "euro money europe currency"],
+  ["💷", "pound money uk sterling currency"],
+  ["💳", "credit card payment debit bank"],
+  ["💸", "spending flying money expense cost"],
+  ["🏦", "bank building finance savings"],
+  ["📈", "chart up growth stocks invest profit"],
+  ["📉", "chart down loss decline stocks"],
+  ["💹", "chart yen growth finance market"],
+  ["🪙", "coin money token currency"],
+  ["🧾", "receipt bill invoice payment"],
+  ["💲", "dollar sign money price cost"],
+  ["🏧", "atm cash withdraw bank machine"],
+  ["🤑", "money face rich greedy finance"],
+  // 🍔 Food & Drink
+  ["🍔", "burger hamburger food fast meal"],
+  ["🍕", "pizza food italian slice"],
+  ["🍣", "sushi food japanese fish"],
+  ["🥗", "salad food healthy green vegetables"],
+  ["🍝", "pasta food italian spaghetti noodles"],
+  ["🌮", "taco food mexican"],
+  ["🌯", "burrito wrap food mexican"],
+  ["🥐", "croissant bakery bread pastry breakfast"],
+  ["🍞", "bread food bakery toast"],
+  ["🥩", "meat steak food beef"],
+  ["🍗", "chicken food poultry drumstick"],
+  ["🍖", "meat bone food ribs"],
+  ["🥓", "bacon food breakfast meat"],
+  ["🍳", "egg fried food breakfast cooking"],
+  ["🥚", "egg food breakfast"],
+  ["🧀", "cheese food dairy"],
+  ["🥪", "sandwich food lunch bread"],
+  ["🌭", "hotdog food fast sausage"],
+  ["🥙", "pita falafel food wrap"],
+  ["🧆", "falafel food middle east"],
+  ["🥘", "stew food pot cooking"],
+  ["🍜", "noodles ramen soup food asian"],
+  ["🍲", "soup stew pot food"],
+  ["🍛", "curry food indian rice"],
+  ["🍱", "bento box food japanese lunch"],
+  ["🥟", "dumpling food asian"],
+  ["🍰", "cake dessert sweet birthday"],
+  ["🍩", "donut doughnut dessert sweet snack"],
+  ["🍪", "cookie biscuit dessert sweet snack"],
+  ["🍫", "chocolate candy sweet snack"],
+  ["🍬", "candy sweet sugar treat"],
+  ["🍭", "lollipop candy sweet"],
+  ["🍿", "popcorn snack cinema movies"],
+  ["🧁", "cupcake dessert sweet bakery"],
+  ["🥧", "pie dessert food bake"],
+  ["🍦", "ice cream dessert cold sweet"],
+  ["🍨", "sundae ice cream dessert"],
+  ["🍧", "shaved ice dessert cold"],
+  ["🥤", "drink cup soda juice beverage"],
+  ["☕", "coffee drink cafe espresso latte"],
+  ["🍵", "tea drink green matcha cup"],
+  ["🫖", "teapot tea drink kettle"],
+  ["🍷", "wine drink alcohol glass bar"],
+  ["🍺", "beer drink alcohol pub bar"],
+  ["🍻", "cheers beer drink celebration"],
+  ["🥂", "champagne toast celebration drink wine"],
+  ["🍸", "cocktail martini drink alcohol bar"],
+  ["🍹", "tropical drink cocktail alcohol"],
+  ["🧃", "juice box drink beverage kids"],
+  ["🥛", "milk drink dairy glass"],
+  ["🍽️", "restaurant dining plate food meal"],
+  ["🧑‍🍳", "chef cook food kitchen restaurant"],
+  ["🫒", "olive food oil mediterranean"],
+  ["🥑", "avocado food healthy green"],
+  ["🍅", "tomato food vegetable red"],
+  ["🥕", "carrot food vegetable orange"],
+  ["🌽", "corn food vegetable yellow"],
+  ["🥦", "broccoli food vegetable green healthy"],
+  ["🧄", "garlic food cooking spice"],
+  ["🧅", "onion food cooking vegetable"],
+  ["🥔", "potato food vegetable"],
+  ["🍠", "sweet potato food vegetable"],
+  ["🍎", "apple fruit food red healthy"],
+  ["🍊", "orange fruit food citrus"],
+  ["🍋", "lemon fruit food citrus yellow"],
+  ["🍌", "banana fruit food yellow"],
+  ["🍇", "grapes fruit food wine purple"],
+  ["🍓", "strawberry fruit food berry red"],
+  ["🫐", "blueberry fruit food berry"],
+  ["🍑", "peach fruit food"],
+  ["🍒", "cherry fruit food red"],
+  ["🥝", "kiwi fruit food green"],
+  ["🍍", "pineapple fruit food tropical"],
+  ["🥭", "mango fruit food tropical"],
+  ["🥥", "coconut fruit food tropical"],
+  ["🌶️", "pepper chili hot spicy food"],
+  // 🏠 Home & Housing
+  ["🏠", "home house housing rent mortgage"],
+  ["🏡", "house garden home property"],
+  ["🏢", "office building work commercial"],
+  ["🏗️", "construction building renovation work"],
+  ["🔑", "key home lock rent property access password security"],
+  ["🛋️", "couch sofa furniture living room"],
+  ["🛏️", "bed bedroom furniture sleep"],
+  ["🚿", "shower bathroom water hygiene"],
+  ["🛁", "bathtub bath bathroom"],
+  ["🪞", "mirror bathroom vanity reflection"],
+  ["🧹", "broom cleaning housework sweep"],
+  ["🧺", "laundry basket cleaning clothes"],
+  ["🪴", "plant pot garden home decor"],
+  ["💡", "lightbulb electricity idea utility"],
+  ["🔌", "plug electricity power utility"],
+  ["🛠️", "tools repair fix maintenance home"],
+  ["🔧", "wrench tool repair fix plumbing"],
+  ["🪛", "screwdriver tool repair fix"],
+  ["🧰", "toolbox repair maintenance handyman"],
+  ["🪣", "bucket cleaning water household"],
+  ["🧴", "lotion bottle hygiene bathroom soap"],
+  ["🧽", "sponge cleaning kitchen household"],
+  ["🪥", "toothbrush hygiene bathroom dental"],
+  ["🚰", "water tap faucet plumbing utility"],
+  ["🪑", "chair furniture seat home"],
+  ["🚪", "door entrance home entry"],
+  ["🪟", "window home glass frame"],
+  ["🛒", "shopping cart grocery store supermarket"],
+  ["🏘️", "houses neighborhood community homes"],
+  ["🧲", "magnet tool household attract science physics"],
+  ["🪤", "mousetrap pest control home"],
+  ["🧯", "fire extinguisher safety emergency"],
+  ["🪜", "ladder tool climb home fix"],
+  ["🖼️", "frame picture art decor wall"],
+  ["🪵", "wood log fireplace material"],
+  ["🧱", "brick building construction material"],
+  // 🚗 Transport
+  ["🚗", "car vehicle automobile transport drive"],
+  ["🚕", "taxi cab ride transport"],
+  ["🚌", "bus public transport commute"],
+  ["🚇", "metro subway train underground transport"],
+  ["🚆", "train rail transport commute travel"],
+  ["🚂", "locomotive train steam railway"],
+  ["🚃", "railway car train carriage"],
+  ["✈️", "airplane flight travel airport plane"],
+  ["🛫", "departure flight airplane takeoff travel"],
+  ["🛬", "arrival flight airplane landing travel"],
+  ["🚢", "ship boat cruise sea transport"],
+  ["⛵", "sailboat boat sailing wind"],
+  ["🚤", "speedboat boat fast water"],
+  ["🛶", "canoe boat paddle kayak water"],
+  ["🛵", "scooter motorcycle transport"],
+  ["🏍️", "motorcycle motorbike transport speed"],
+  ["🚲", "bicycle bike cycling transport"],
+  ["🛴", "scooter kick transport"],
+  ["🛺", "rickshaw auto transport tuk"],
+  ["⛽", "fuel gas petrol station car"],
+  ["🅿️", "parking car park"],
+  ["🚦", "traffic light signal road"],
+  ["🛤️", "railway tracks train"],
+  ["🚁", "helicopter flight transport"],
+  ["🛞", "tire wheel car rubber"],
+  ["🚘", "car oncoming vehicle drive"],
+  ["🚙", "suv car vehicle sport"],
+  ["🛻", "pickup truck vehicle transport"],
+  ["🚚", "delivery truck transport moving"],
+  ["🚛", "semi truck transport freight"],
+  ["🚐", "minibus van transport"],
+  ["🚎", "trolleybus transport electric public"],
+  ["🛣️", "highway road motorway freeway"],
+  // 🛍️ Shopping & Clothing
+  ["🛍️", "shopping bags mall store retail"],
+  ["👕", "shirt tshirt clothes clothing fashion"],
+  ["👗", "dress clothes clothing fashion women"],
+  ["👟", "sneakers shoes running sports footwear"],
+  ["👠", "heels shoes fashion women footwear"],
+  ["👜", "handbag purse bag fashion accessories"],
+  ["🧥", "coat jacket clothes winter warm"],
+  ["👔", "tie necktie formal work clothes"],
+  ["👒", "hat fashion sun accessories"],
+  ["🧢", "cap hat fashion accessories"],
+  ["💎", "gem diamond jewelry luxury"],
+  ["⌚", "watch time accessories luxury fashion"],
+  ["👓", "glasses eyewear optical vision"],
+  ["🕶️", "sunglasses fashion accessories cool"],
+  ["💄", "lipstick makeup cosmetics beauty"],
+  ["👙", "bikini swimsuit beach summer clothes"],
+  ["🩱", "swimsuit one piece bathing clothes"],
+  ["👘", "kimono japanese clothes traditional"],
+  ["🥻", "sari clothes traditional indian"],
+  ["🧣", "scarf winter warm clothes accessory"],
+  ["🧤", "gloves winter warm hands clothes"],
+  ["🧦", "socks clothes feet warm"],
+  ["👞", "shoe formal dress leather footwear"],
+  ["👡", "sandal shoe summer open footwear"],
+  ["🥾", "hiking boot shoe outdoor footwear"],
+  ["👑", "crown king queen royal luxury"],
+  ["🎀", "ribbon bow gift decoration"],
+  ["👝", "clutch purse bag evening"],
+  ["🎽", "running shirt sport athletic clothes"],
+  ["🩳", "shorts clothes summer casual"],
+  ["🩴", "flip flop sandal summer beach"],
+  ["💅", "nail polish manicure beauty fashion"],
+  // 🏥 Health & Wellness
+  ["🏥", "hospital health medical doctor emergency"],
+  ["💊", "pill medicine medication pharmacy health"],
+  ["💉", "syringe vaccine injection medical"],
+  ["🩺", "stethoscope doctor medical checkup"],
+  ["🩹", "bandage first aid medical injury"],
+  ["🦷", "tooth dental dentist health"],
+  ["👁️", "eye vision optical glasses health"],
+  ["🧘", "yoga meditation wellness mindfulness"],
+  ["🏋️", "gym weights fitness workout exercise"],
+  ["🤸", "gymnastics exercise sport fitness"],
+  ["💆", "massage spa relax wellness"],
+  ["🧖", "spa sauna steam wellness relax"],
+  ["❤️", "heart love health red"],
+  ["🫀", "heart organ medical health"],
+  ["🧠", "brain mind mental health thinking"],
+  ["♿", "disability accessibility wheelchair"],
+  ["🩻", "xray bones medical scan"],
+  ["🩼", "crutch injury support medical"],
+  ["🦴", "bone skeleton medical body"],
+  ["🫁", "lungs breathing medical organ"],
+  ["🩸", "blood drop medical donate"],
+  ["🏃", "running exercise cardio fitness jog"],
+  ["🚴", "cycling bike exercise sport"],
+  ["🏊", "swimming pool exercise sport water"],
+  ["🧗", "climbing sport exercise adventure"],
+  ["🤾", "handball sport exercise ball"],
+  ["🤺", "fencing sport exercise sword"],
+  ["🏄", "surfing sport water wave beach"],
+  ["⚕️", "medical health caduceus symbol"],
+  ["🧬", "dna genetics science biology strand"],
+  ["🔬", "microscope science research lab medical"],
+  // 🎓 Education & Work
+  ["🎓", "graduation education university school degree"],
+  ["📚", "books reading education library study"],
+  ["📖", "book reading education study open"],
+  ["✏️", "pencil writing education draw"],
+  ["📝", "memo notes writing study document"],
+  ["🖥️", "computer desktop screen work tech monitor"],
+  ["💼", "briefcase business work office job"],
+  ["🏫", "school education building class"],
+  ["📐", "ruler triangle math geometry education"],
+  ["🧪", "test tube science lab chemistry"],
+  ["📏", "ruler measure straight edge office"],
+  ["🎒", "backpack school bag student education"],
+  ["📓", "notebook journal writing study"],
+  ["🗂️", "file folder organize office document"],
+  ["📋", "clipboard list check document"],
+  ["🖊️", "pen writing ink office"],
+  ["🖋️", "fountain pen writing calligraphy"],
+  ["✒️", "nib pen ink writing"],
+  ["📎", "paperclip office supply attach"],
+  ["🗃️", "card file box office organize"],
+  ["📊", "bar chart statistics data analytics"],
+  ["📑", "tabs document bookmark page"],
+  ["📰", "newspaper news media press read"],
+  ["🗞️", "rolled newspaper news media press"],
+  ["📒", "ledger notebook yellow pages"],
+  ["📕", "book closed red reading"],
+  ["📗", "book green reading textbook"],
+  ["📘", "book blue reading textbook"],
+  ["📙", "book orange reading textbook"],
+  ["🧑‍💻", "programmer developer coder tech work"],
+  ["🧑‍🏫", "teacher education instructor class"],
+  ["🧑‍🔬", "scientist research lab education"],
+  ["🧑‍⚕️", "doctor health medical professional"],
+  ["🧑‍🔧", "mechanic repair fix tool worker"],
+  ["🧑‍🌾", "farmer agriculture garden food"],
+  ["🧑‍🎨", "artist creative paint design"],
+  ["🧑‍✈️", "pilot aviation fly captain"],
+  ["🧑‍🚒", "firefighter rescue emergency safety"],
+  ["👷", "construction worker build hard hat"],
+  ["👮", "police officer law enforcement cop"],
+  // 🎮 Entertainment & Fun
+  ["🎮", "gaming controller video games play"],
+  ["🎬", "movie cinema film clapper"],
+  ["🎭", "theater drama performing arts masks"],
+  ["🎤", "microphone karaoke sing music"],
+  ["🎧", "headphones music audio listen"],
+  ["🎵", "music note song sound"],
+  ["🎶", "music notes song melody"],
+  ["🎸", "guitar music rock instrument"],
+  ["🎹", "piano keyboard music instrument"],
+  ["🎺", "trumpet music brass instrument"],
+  ["🥁", "drum music instrument beat"],
+  ["🎲", "dice game board luck casino"],
+  ["🧩", "puzzle game jigsaw piece"],
+  ["🎯", "target bullseye goal darts aim"],
+  ["🎪", "circus tent carnival fun"],
+  ["🎠", "carousel ride amusement park fun"],
+  ["🎰", "slot machine casino gambling jackpot"],
+  ["🎱", "billiards pool ball game"],
+  ["🕹️", "joystick arcade retro game"],
+  ["🎷", "saxophone music jazz instrument"],
+  ["🪕", "banjo music instrument country"],
+  ["🎻", "violin fiddle music instrument classical"],
+  ["📻", "radio music audio broadcast"],
+  ["🎙️", "studio microphone mic podcast recording audio"],
+  ["📀", "dvd disc movie media"],
+  ["💿", "cd disc music album"],
+  ["🎞️", "film frames movie cinema reel"],
+  ["📸", "camera flash photo selfie"],
+  ["🎨", "art palette paint creative drawing"],
+  ["🖌️", "paintbrush art creative drawing"],
+  ["🖍️", "crayon draw art color kids"],
+  ["🎳", "bowling sport ball pins game"],
+  ["🏆", "trophy winner champion award prize"],
+  ["🥇", "gold medal first winner champion"],
+  ["🥈", "silver medal second runner up"],
+  ["🥉", "bronze medal third place"],
+  ["🏅", "medal award sports achievement winner"],
+  // ⚽ Sports & Activities
+  ["⚽", "soccer football sport ball kick"],
+  ["🏀", "basketball sport ball hoop"],
+  ["🏈", "football american sport ball"],
+  ["⚾", "baseball sport ball bat"],
+  ["🥎", "softball sport ball"],
+  ["🎾", "tennis sport ball racket"],
+  ["🏐", "volleyball sport ball net"],
+  ["🏉", "rugby sport ball"],
+  ["🏸", "badminton sport shuttlecock racket"],
+  ["🏓", "table tennis ping pong sport paddle"],
+  ["🥊", "boxing glove sport fight"],
+  ["🥋", "martial arts karate judo sport"],
+  ["⛳", "golf sport flag hole"],
+  ["⛷️", "skiing sport winter snow"],
+  ["🏂", "snowboard sport winter snow"],
+  ["🛷", "sled sport winter snow toboggan"],
+  ["⛸️", "ice skating sport winter"],
+  ["🏌️", "golf sport swing club"],
+  ["🎿", "ski sport winter snow poles"],
+  ["🛹", "skateboard sport ride trick"],
+  ["🤿", "diving snorkel sport underwater"],
+  ["🏹", "archery bow arrow sport"],
+  // 🏖️ Travel & Nature
+  ["🌴", "palm tree tropical vacation beach"],
+  ["🏖️", "beach vacation sand sun summer"],
+  ["⛱️", "umbrella beach sun shade summer"],
+  ["🌍", "world globe earth travel map"],
+  ["🌎", "world globe americas travel"],
+  ["🌏", "world globe asia travel"],
+  ["🗺️", "map world travel navigation"],
+  ["🏔️", "mountain snow nature hiking"],
+  ["⛺", "tent camping outdoor nature"],
+  ["🏕️", "camping outdoor tent nature park"],
+  ["🌅", "sunrise sunset morning nature"],
+  ["🌊", "wave ocean sea water surf"],
+  ["🏝️", "island tropical vacation beach"],
+  ["🗼", "tower landmark tourism sightseeing"],
+  ["🗽", "statue liberty usa new york landmark"],
+  ["🎢", "roller coaster amusement park fun ride"],
+  ["🏰", "castle kingdom fairy tale landmark"],
+  ["⛩️", "shrine temple japan torii gate"],
+  ["🕌", "mosque religion islam prayer"],
+  ["⛪", "church religion christian prayer"],
+  ["🕍", "synagogue religion jewish prayer temple"],
+  ["🛕", "hindu temple religion prayer"],
+  ["🕋", "kaaba mecca islam holy"],
+  ["🌋", "volcano nature mountain hot lava"],
+  ["🏜️", "desert sand dry hot nature"],
+  ["🌲", "evergreen tree forest nature pine"],
+  ["🌳", "tree nature deciduous forest"],
+  ["🌸", "cherry blossom flower spring japan"],
+  ["🌺", "hibiscus flower tropical"],
+  ["🌻", "sunflower flower yellow nature"],
+  ["🌹", "rose flower red love romantic"],
+  ["🌷", "tulip flower spring garden"],
+  ["💐", "bouquet flowers gift arrangement"],
+  ["🌿", "herb leaf green plant nature"],
+  ["🍀", "four leaf clover luck irish"],
+  ["🍁", "maple leaf fall autumn canada"],
+  ["🍂", "fallen leaf autumn nature"],
+  ["🌾", "rice ear grain crop farm"],
+  ["🏞️", "national park nature landscape valley"],
+  ["🌌", "milky way galaxy night sky stars"],
+  ["🌈", "rainbow nature weather colors"],
+  ["🌧️", "rain cloud weather water"],
+  ["⛈️", "thunder storm lightning weather"],
+  ["🌪️", "tornado storm wind weather"],
+  ["🌤️", "partly cloudy sun weather"],
+  ["🌙", "moon crescent night"],
+  ["🌕", "full moon night sky"],
+  // 🐕 Pets & Animals
+  ["🐕", "dog pet puppy animal canine"],
+  ["🐱", "cat pet kitten animal feline"],
+  ["🐟", "fish pet aquarium animal sea"],
+  ["🐠", "tropical fish pet aquarium colorful"],
+  ["🐡", "blowfish puffer fish sea animal"],
+  ["🐦", "bird pet animal flying sing"],
+  ["🦜", "parrot bird pet colorful talking"],
+  ["🦅", "eagle bird animal predator"],
+  ["🐹", "hamster pet small animal cute"],
+  ["🐰", "rabbit bunny pet animal cute"],
+  ["🐢", "turtle pet animal slow reptile"],
+  ["🐾", "paw print animal pet foot"],
+  ["🦮", "guide dog service pet animal"],
+  ["🐈", "cat pet animal feline kitty"],
+  ["🐎", "horse animal ride equestrian"],
+  ["🦊", "fox animal wild clever"],
+  ["🐻", "bear animal wild nature"],
+  ["🐼", "panda bear animal cute china"],
+  ["🦁", "lion animal wild king jungle"],
+  ["🐯", "tiger animal wild stripes"],
+  ["🐮", "cow animal farm dairy milk"],
+  ["🐷", "pig animal farm pink"],
+  ["🐑", "sheep animal farm wool lamb"],
+  ["🐔", "chicken animal farm hen"],
+  ["🦆", "duck animal bird water"],
+  ["🦢", "swan bird animal elegant white"],
+  ["🐍", "snake reptile animal slither"],
+  ["🦎", "lizard reptile animal gecko"],
+  ["🐊", "crocodile alligator animal reptile"],
+  ["🐘", "elephant animal large trunk"],
+  ["🦒", "giraffe animal tall spots"],
+  ["🐪", "camel animal desert hump"],
+  ["🐧", "penguin animal bird cold arctic"],
+  ["🦋", "butterfly insect nature colorful"],
+  ["🐝", "bee insect honey buzz pollinate"],
+  ["🐛", "bug insect caterpillar worm"],
+  ["🕷️", "spider insect web arachnid"],
+  ["🐌", "snail slow shell garden"],
+  ["🐬", "dolphin marine animal sea smart"],
+  ["🐳", "whale marine animal sea large"],
+  ["🦈", "shark marine animal sea predator"],
+  ["🐙", "octopus marine animal sea tentacle"],
+  // 👶 Family & People
+  ["👶", "baby infant child family newborn"],
+  ["👧", "girl child daughter kid"],
+  ["👦", "boy child son kid"],
+  ["👨‍👩‍👧", "family parents child household"],
+  ["👪", "family parents children household"],
+  ["👨‍👩‍👧‍👦", "family four parents kids"],
+  ["🤰", "pregnant maternity baby expecting"],
+  ["🎂", "birthday cake celebration party"],
+  ["🎁", "gift present birthday holiday surprise"],
+  ["💍", "ring engagement wedding marriage jewelry"],
+  ["💒", "wedding chapel marriage ceremony"],
+  ["💏", "couple kiss love romance"],
+  ["👰", "bride wedding marriage ceremony"],
+  ["🤵", "groom wedding marriage formal"],
+  ["👩‍❤️‍👨", "couple love relationship heart"],
+  ["🧓", "elder old senior grandparent"],
+  ["👴", "grandpa grandfather old man senior"],
+  ["👵", "grandma grandmother old woman senior"],
+  ["🧒", "child kid young person"],
+  ["🎉", "party celebration confetti birthday fun"],
+  ["🎊", "confetti ball celebration party"],
+  ["🥳", "party face celebration birthday hat"],
+  ["🎈", "balloon party celebration decoration"],
+  ["🎄", "christmas tree holiday xmas december"],
+  ["🎃", "pumpkin halloween jack lantern october"],
+  ["🕎", "menorah hanukkah jewish holiday candle"],
+  ["🪔", "diya lamp light festival diwali"],
+  ["🧧", "red envelope chinese new year lucky"],
+  // 📱 Tech & Communication
+  ["📱", "phone mobile smartphone cell device"],
+  ["💻", "laptop computer device tech work"],
+  ["🖨️", "printer print office document tech"],
+  ["📞", "telephone phone call communication"],
+  ["📶", "signal wifi internet bars connection"],
+  ["📡", "satellite antenna signal communication"],
+  ["🔋", "battery power charge energy device"],
+  ["💾", "floppy disk save storage data"],
+  ["📷", "camera photo photography picture"],
+  ["🎥", "video camera film record"],
+  ["📺", "television tv screen watch"],
+  ["🖱️", "mouse computer click device"],
+  ["⌨️", "keyboard typing computer device"],
+  ["🖲️", "trackball computer input device"],
+  ["💽", "minidisc storage data media"],
+  ["📼", "vhs tape video cassette media"],
+  ["📹", "camcorder video record camera"],
+  ["🔍", "magnifying glass search find zoom"],
+  ["🔎", "magnifying glass search find right"],
+  ["🌐", "globe internet web world network"],
+  ["📧", "email envelope message inbox"],
+  ["📩", "inbox envelope message receive"],
+  ["📨", "incoming envelope message mail"],
+  ["💬", "speech bubble chat message talk"],
+  ["📲", "mobile arrow phone incoming call"],
+  ["📟", "pager beeper device old"],
+  ["🔊", "speaker volume loud audio sound"],
+  ["🔈", "speaker low volume audio sound"],
+  // 🏛️ Government & Services
+  ["🏛️", "government building official institution"],
+  ["⚖️", "scales justice law legal court"],
+  ["📜", "scroll document law certificate"],
+  ["🗳️", "ballot vote election democracy"],
+  ["🚔", "police car law enforcement"],
+  ["🚒", "fire truck emergency rescue"],
+  ["🚑", "ambulance emergency medical hospital"],
+  ["📮", "mailbox post letter mail"],
+  ["🏣", "post office mail service"],
+  ["🏤", "european post office building mail"],
+  ["🏪", "convenience store shop 24 open"],
+  ["🏬", "department store shopping mall"],
+  ["🏨", "hotel accommodation sleep travel"],
+  ["🏩", "love hotel accommodation romance"],
+  ["🏟️", "stadium arena sport event"],
+  ["🏙️", "cityscape buildings urban skyline"],
+  ["🌃", "night city buildings urban skyline"],
+  ["🌆", "sunset cityscape buildings evening"],
+  ["🛃", "customs border passport immigration"],
+  ["🛂", "passport control immigration border"],
+  // ⚡ Utilities & Symbols
+  ["⚡", "lightning electricity power energy bolt"],
+  ["💧", "water drop utility plumbing"],
+  ["🔥", "fire hot flame gas heating"],
+  ["♻️", "recycle green environment eco"],
+  ["🌱", "seedling plant grow garden nature"],
+  ["☀️", "sun solar sunshine weather energy"],
+  ["❄️", "snowflake cold winter freeze ice"],
+  ["🌡️", "thermometer temperature weather heat"],
+  ["🚫", "prohibited no block forbidden ignore"],
+  ["⭐", "star favorite important gold"],
+  ["✅", "check mark done complete approved"],
+  ["❌", "cross reject cancel remove delete"],
+  ["🔔", "bell notification alert reminder"],
+  ["🔒", "lock security password private safe"],
+  ["📌", "pin pushpin location bookmark mark"],
+  ["🏷️", "tag label price sale category"],
+  ["❓", "question mark help unknown what"],
+  ["❗", "exclamation mark important alert warning"],
+  ["⚠️", "warning caution danger alert triangle"],
+  ["💯", "hundred percent score perfect"],
+  ["🗑️", "trash can delete garbage waste bin"],
+  ["✂️", "scissors cut trim tool"],
+  ["📦", "package box shipping delivery parcel"],
+  ["🔗", "link chain url connection"],
+  ["🏴", "black flag pirate skull"],
+  ["🏳️", "white flag surrender peace"],
+  ["🇮🇱", "israel flag country"],
+  ["🇺🇸", "usa america flag country"],
+  ["🇬🇧", "uk britain flag country england"],
+  ["🇪🇺", "european union flag eu"],
+  ["🔄", "arrows cycle refresh reload update"],
+  ["⏰", "alarm clock time morning wake"],
+  ["⏱️", "stopwatch timer time speed"],
+  ["🕐", "clock one time oclock hour"],
+  ["📆", "calendar date schedule plan month"],
+  ["📅", "calendar date event schedule"],
+  ["🗓️", "calendar spiral date schedule plan"],
+  ["✨", "sparkles magic new shiny clean"],
+  ["💫", "dizzy star shooting sparkle"],
+  ["🎗️", "ribbon awareness cancer support"],
+  ["🔮", "crystal ball fortune magic predict"],
+  ["🧿", "evil eye nazar blue protection"],
+  ["🪬", "hamsa hand protection luck"],
+  ["♠️", "spade card suit game black"],
+  ["♥️", "heart card suit game red"],
+  ["♦️", "diamond card suit game red"],
+  ["♣️", "club card suit game black"],
+  ["🀄", "mahjong game tile chinese"],
+  ["🎴", "hanafuda card game japanese"],
+  // 🏠 Real Estate & Property
+  ["🏚️", "derelict house abandoned building old"],
+  ["🪨", "rock stone nature boulder"],
+  ["⛲", "fountain water park garden"],
+  ["🌉", "bridge night city travel structure"],
+  ["🌁", "foggy bridge san francisco city"],
+  // 🧸 Kids & Toys
+  ["🧸", "teddy bear toy stuffed animal kids plush cute"],
+  ["🪀", "yo-yo toy play kids"],
+  ["🪁", "kite toy play wind outdoor kids"],
+  ["🎎", "dolls japanese festival decoration"],
+  ["🪆", "nesting dolls russian matryoshka toy"],
+  ["🎐", "wind chime decoration bell sound"],
+  // 📐 Math & Science
+  ["🧮", "abacus counting math calculate"],
+  ["🔢", "numbers input 1234 math"],
+  ["🔣", "symbols input special characters"],
+  ["🧫", "petri dish science biology lab"],
+  ["🔭", "telescope astronomy space star science"],
+  ["⚗️", "alembic chemistry science distill lab"],
+  ["⚛️", "atom physics science nuclear"],
+];
